Fixes all remaining (relevant) todos and adds proper error pages.
