+++ Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 askama = { version = "0.12.0", features = ["with-axum"] }
 askama_axum = "0.3.0"
 axum = "0.6.20"
+rust-embed = { version = "8.0.0", features = ["axum"] }
 chrono = "0.4.31"
 csv = "1.2.2"
 dotenvy = "0.15.7"
@@ -20,8 +21,8 @@ reqwest = { version = "0.11.20", features = ["stream"] }
 serde = { version = "1.0.188", features = ["derive"] }
 serde_json = "1.0.107"
 sqlx = { version = "0.7.2", features = ["sqlite", "runtime-tokio"] }
-static_dir = "0.2.0"
 tempfile = "3.8.0"
 tokio = { version = "1.32.0", features = ["rt-multi-thread", "macros"] }
 zstd = "0.12.4"
 tower-http = { version = "0.4.4", features = ["fs"] }
+mime_guess = "2.0.4"
