+++ templates/header.html
@@ -17,7 +17,7 @@
 <nav class="navbar" role="navigation" aria-label="main navigation">
     <div id="nav-menu" class="my-navbar-menu">
         <a class="navbar-item" href="/">
-            Home
+            Stats
         </a>
 
         <a class="navbar-item" href="/tactics">
@@ -25,7 +25,7 @@
         </a>
 
         <a class="navbar-item" href="/tactics/new">
-            Practice
+            Next Puzzle
         </a>
 
         <a class="navbar-item" href="/about">
