+++ templates/index.html
@@ -34,7 +34,7 @@
             </tbody>
         </table>
         <a href="/tactics" class="button">Review</a>
-        <a href="/tactics/new" class="button">New Puzzle</a>
+        <a href="/tactics/new" class="button">Next Puzzle</a>
     </div>
 
     <!-- The review forecast -->
