+++ templates/not_found.html
@@ -1,0 +1,16 @@
+{% extends "base.html" %}
+
+{% block content %}
+<div class="columns">
+    <div id="about" class="column bt-panel">
+        <h2 class="title is-3">
+            Not found
+        </h2>
+        <p>
+            The requested page {{request_uri}} was not found. It's possible that the address of the
+            requested page has changed, or that it no longer exists. Try going to the
+            <a href="/">main page</a> or finding what you're looking for on the menu in the top right.
+        </p>
+    </div>
+</div>
+{% endblock %}
